Now using webpack for all-the-things!
