+++ .gitignore
@@ -6,4 +6,3 @@
 
 # FILES
 .DS_Store
-typer-es5.js
