+++ less/typer.less
@@ -29,6 +29,10 @@
   content: none;
 }
 
+.white-space {
+  white-space: pre-wrap;
+}
+
 @keyframes softblink {
   0% { opacity: 1; }
   50% { opacity: 0; }
@@ -40,8 +44,4 @@
   49% { opacity: 1; }
   50% { opacity: 0; }
   100% { opacity: 0; }
-}
-
-.white-space {
-  white-space: pre-wrap;
 }
