+++ postcss.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  plugins: {
+    'autoprefixer': {},
+    cssnano: {}
+  }
+}
