+++ typer.css
@@ -1,53 +1,1 @@
-.typer {
-  position: relative;
-  min-height: 2ex;
-}
-.typer::after {
-  content: '';
-  position: absolute;
-  display: inline-block;
-  width: 0.1em;
-  height: 1.1em;
-  background-color: currentColor;
-  margin-left: 0.1em;
-}
-.cursor-block::after {
-  width: 0.9ex;
-}
-.cursor-soft::after {
-  animation: softblink 0.7s infinite;
-}
-.cursor-hard::after {
-  animation: hardblink 0.7s infinite;
-}
-.no-cursor::after {
-  content: none;
-}
-@keyframes softblink {
-  0% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-@keyframes hardblink {
-  0% {
-    opacity: 1;
-  }
-  49% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 0;
-  }
-}
-.white-space {
-  white-space: pre-wrap;
-}
+.typer{min-height:2ex;position:relative}.typer:after{background-color:currentColor;content:"";display:inline-block;height:1.1em;margin-left:.1em;position:absolute;width:.1em}.cursor-block:after{width:.9ex}.cursor-soft:after{-webkit-animation:softblink .7s infinite;animation:softblink .7s infinite}.cursor-hard:after{-webkit-animation:hardblink .7s infinite;animation:hardblink .7s infinite}.no-cursor:after{content:none}.white-space{white-space:pre-wrap}@-webkit-keyframes softblink{0%{opacity:1}50%{opacity:0}to{opacity:1}}@keyframes softblink{0%{opacity:1}50%{opacity:0}to{opacity:1}}@-webkit-keyframes hardblink{0%{opacity:1}49%{opacity:1}50%{opacity:0}to{opacity:0}}@keyframes hardblink{0%{opacity:1}49%{opacity:1}50%{opacity:0}to{opacity:0}}
